+++ src/main/java/cn/lefer/tomu/mapper/PlayHistoryMapper.java
@@ -18,5 +18,5 @@ import java.util.Date;
 public interface PlayHistoryMapper {
     int insert(PlayHistory playHistory);
     PlayHistory selectPlayStatusByChannelID(@Param("channelID") int channelID);
-    int updateStatus(@Param("lastPosition") int lastPosition,@Param("playDate") Date playDate,@Param("playHistoryID") int playHistoryID);
+    int updateStatus(@Param("lastPosition") double lastPosition,@Param("playDate") Date playDate,@Param("playHistoryID") int playHistoryID);
 }
